+++ src/bin/server.rs
@@ -51,10 +51,11 @@ async fn main() -> io::Result<()> {
 
             .route("/", get().to(pages::home::get))
             .route("/about", get().to(pages::about::get))
-            .route("/images", post().to(handlers::images::upload))
             .route("/sandbox", get().to(pages::sandbox::get))
             .service(pages::auth::resource("/auth"))
             .service(pages::posts::resource("/posts"))
+            .service(pages::admin::resource("/admin"))
+            .route("/api/images", post().to(handlers::images::upload))
     })
     .bind("127.0.0.1:3030")?
     .run()
