+++ src/extensions.rs
@@ -2,5 +2,5 @@ use crate::models::AuthorWithoutPassword;
 
 #[derive(Debug)]
 pub struct Assigns {
-    pub user: Option<AuthorWithoutPassword>,
+    pub author: Option<AuthorWithoutPassword>,
 }
