+++ src/middleware.rs
@@ -10,7 +10,7 @@ pub fn set_assigns<S, B>(req: ServiceRequest, service: &mut S) -> S::Future
 where
     S: Service<Request = ServiceRequest, Response = ServiceResponse<B>, Error = Error>
 {
-    req.head().extensions_mut().insert(Assigns{ user: None });
+    req.head().extensions_mut().insert(Assigns{ author: None });
     service.call(req)
 }
 
@@ -19,9 +19,13 @@ pub fn load_user<S, B>(req: ServiceRequest, service: &mut S) -> S::Future
 where
     S: Service<Request = ServiceRequest, Response = ServiceResponse<B>, Error = Error>
 {
-    if let Ok(Some(author)) = req.get_session().get::<AuthorWithoutPassword>("auth") {
-        if let Some(assigns) = req.head().extensions_mut().get_mut::<Assigns>() {
-            assigns.user = Some(author);
+    {
+        let auth = req.get_session().get::<AuthorWithoutPassword>("auth");
+        let mut exts = req.head().extensions_mut();
+        let assigns = exts.get_mut::<Assigns>();
+
+        if let (Ok(auth), Some(assigns)) = (auth, assigns) {
+            assigns.author = auth;
         }
     }
