+++ src/pages/mod.rs
@@ -1,6 +1,5 @@
-use actix_web::HttpResponse;
-
 pub mod about;
+pub mod admin;
 pub mod auth;
 pub mod home;
 pub mod notfound;
@@ -10,10 +9,3 @@ pub mod sandbox;
 mod page;
 
 pub use page::Page;
-
-
-pub trait Renderable: Into<String> {
-    fn render(self) -> HttpResponse {
-        HttpResponse::Ok().body::<String>(self.into())
-    }
-}
